auth ability, login

+++ app/Http/Controllers/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Task;
 use App\Models\User;
 
@@ -11,6 +12,15 @@ class TaskController extends Controller
 {
     public function index(){
         $tasks = Task::with(['user'])
+                    ->where('confidential', false)
+                    ->orderByDesc('id')
+                    ->get();
+        return $tasks;
+    }
+
+    public function confidentialTasks(){
+        $tasks = Task::with(['user'])
+                    ->where('confidential', true)
                     ->orderByDesc('id')
                     ->get();
         return $tasks;
@@ -18,17 +28,27 @@ class TaskController extends Controller
 
     public function show($id){
         $task = Task::findOrFail($id);
+        
+        //check if can show confidentail
+        if($task->confidential == true && Auth::user()->rank < 2)
+            return abort(403);
+
         return $task->load(['user']);
     }
 
     public function store(Request $request){
+        //check if rank > 0
+        if(!Auth::user()->checkRank())
+            return abort(403);
+
         $data = $request->validate([
             'name' => 'nullable|string|max:50',
             'content' => 'nullable|string|max:1000',
             'user_id' => [
                 'required',
                 Rule::in(User::query()->pluck('id')->toArray())
-            ]
+            ],
+            'confidential' => 'nullable|boolean'
         ]);
 
         $task = new Task($data);
@@ -40,13 +60,18 @@ class TaskController extends Controller
     }
 
     public function update(Request $request, $id){
+        //check if rank > 0
+        if(!Auth::user()->checkRank())
+            return abort(403);
+
         $data = $request->validate([
             'name' => 'nullable|string|max:50',
             'content' => 'nullable|string|max:1000',
             'user_id' => [
                 'required',
                 Rule::in(User::query()->pluck('id')->toArray())
-            ]
+            ],
+            'confidential' => 'nullable|boolean'
         ]);
 
         $task = Task::findOrFail($id);
@@ -59,6 +84,10 @@ class TaskController extends Controller
     }
 
     public function delete($id){
+        //check if rank > 0
+        if(!Auth::user()->checkRank())
+            return abort(403);
+
         $task = Task::findOrFail($id);
         
         if($task->delete())
